--- mysite/news/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .views import latest_news
-
-urlpatterns = [
-    path('api/news/', latest_news),
-]
