+++ news/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from .views import latest_news
+
+urlpatterns = [
+    path('api/news/', latest_news),
+]
