+++ templates/index.html
@@ -1,0 +1,204 @@
+<!DOCTYPE html>
+<html lang="ko">
+  <head>
+    <meta charset="UTF-8" />
+    <title>뉴스 검색</title>
+    <style>
+      body {
+        font-family: Arial, sans-serif;
+        margin: 0;
+        padding: 0;
+        background: #111;
+        color: #fff;
+      }
+
+      /* 공통 */
+      .center {
+        /* flex 정렬로 가운데 정렬(페이지의 수직, 수평 중앙) */
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+      }
+
+      /* 메인 페이지 */
+      .main-container {
+        /* 메인 화면의 전체 */
+        height: 100vh;
+      }
+
+      .logo {
+        /* yaho, 홈페이지 이름을 보여줄 곳 */
+        font-size: 30px;
+        margin-bottom: 20px;
+        color: #0ff;
+      }
+
+      .search-box {
+        /* 검색어 박스, 버튼의 위치 나란히 정렬 */
+        display: flex;
+      }
+
+      .search-input {
+        /* 검색 입력 스타일 */
+        padding: 10px;
+        font-size: 16px;
+        width: 300px;
+        border: none;
+        border-radius: 5px 0 0 5px;
+      }
+
+      .search-button {
+        padding: 10px 20px;
+        font-size: 16px;
+        border: none;
+        background: #0ff;
+        color: #000;
+        border-radius: 0 5px 5px 0;
+        cursor: pointer;
+      }
+
+      /* 검색 결과 페이지 */
+      .results-container {
+        padding: 30px;
+      }
+
+      .current-keyword {
+        text-align: center;
+        font-size: 24px;
+        margin-bottom: 20px;
+        color: #0ff;
+      }
+
+      .scroll-box {
+        max-height: 400px;
+        overflow-y: auto;
+        border: 1px solid #555;
+        padding: 15px;
+        border-radius: 10px;
+        background: #222;
+      }
+
+      .headline-item {
+        padding: 10px;
+        margin-bottom: 10px;
+        background: #fff;
+        color: #000;
+        border-radius: 5px;
+        transition: background 0.3s, color 0.3s;
+        cursor: pointer;
+      }
+
+      .headline-item:hover {
+        background: #0ff;
+        color: #000;
+        font-weight: bold;
+      }
+
+      /* 뉴스 상세 페이지 */
+      .news-detail-container {
+        display: flex;
+        padding: 30px;
+        gap: 20px;
+      }
+
+      .news-left {
+        width: 60%;
+        background: #222;
+        padding: 20px;
+        border-radius: 10px;
+      }
+
+      .news-right {
+        width: 40%;
+        background: #333;
+        padding: 20px;
+        border-radius: 10px;
+      }
+
+      .news-headline {
+        font-size: 22px;
+        margin-bottom: 15px;
+        color: #0ff;
+      }
+
+      .news-full-text {
+        max-height: 400px;
+        overflow-y: auto;
+        line-height: 1.6;
+      }
+
+      .summary-box {
+        min-height: 100px;
+        background: #444;
+        padding: 10px;
+        border-radius: 5px;
+      }
+
+      .action-buttons button {
+        margin-top: 10px;
+        padding: 10px;
+        background: #0ff;
+        border: none;
+        border-radius: 5px;
+        cursor: pointer;
+        font-weight: bold;
+      }
+
+      .back-button {
+        margin-top: 30px;
+        display: inline-block;
+        padding: 10px 20px;
+        background: #0ff;
+        color: #000;
+        text-decoration: none;
+        border-radius: 5px;
+      }
+    </style>
+  </head>
+
+  <body>
+    <!-- 메인 페이지 -->
+    <div class="main-container center">
+      <div class="logo">yaho</div>
+      <form class="search-box" action="/search/" method="get">
+        <input class="search-input" type="text" name="q" placeholder="검색어를 입력하세요" />
+        <button class="search-button" type="submit">검색</button>
+      </form>
+    </div>
+
+    <!-- 검색 결과 페이지 -->
+    <div class="results-container">
+      <div class="current-keyword">"정치" 검색 결과</div>
+      <form class="search-box center" action="/search/" method="get">
+        <input class="search-input" type="text" name="q" placeholder="다시 검색" />
+        <button class="search-button" type="submit">검색</button>
+      </form>
+      <div class="scroll-box">
+        <div class="headline-item">~헤드라인 1~</div>
+        <div class="headline-item">~헤드라인 2~</div>
+        <div class="headline-item">~헤드라인 3~</div>
+        <!-- 계속 반복 -->
+      </div>
+    </div>
+
+    <!-- 뉴스 상세 페이지 -->
+    <div class="news-detail-container">
+      <div class="news-left">
+        <div class="news-headline">~헤드라인 제목~</div>
+        <div class="news-full-text">~전체 기사 내용~</div>
+      </div>
+      <div class="news-right">
+        <div class="summary-box">
+          <h3>요약</h3>
+          <p>~요약 내용~</p>
+        </div>
+        <div class="action-buttons">
+          <button>요약 읽어주기</button>
+          <button>전문 읽어주기</button>
+        </div>
+        <a href="/search/?q=정치" class="back-button">이전 페이지로</a>
+      </div>
+    </div>
+  </body>
+</html>
